Fixed crash with validationHandler closure
Turns out, the Box subclass was not needed anymore either as Swift 3
allows you to cast anything into AnyObject. This removes our need to
convert the validationHandler closure into an object to store with
Objective-C.

However, the initial crash was caused by what seems to be a bug in how
NSObject handles Generics, although I don’t know the exact nature of
it. By removing the NSObject subclass from Box<T>, we no longer crashed
on validation.

Big thanks to mikeash on #swift-lang for helping me through this!